More conversions to long. I've found out that C# imposes a maximum to array index 2^31. So for byte arrays the max is 2GB. If there's a way around it I'll continue the conversion to long.

+++ FloppyControlApp.net7/WindowsFormsApplication2/WindowsFormsApplication2/MyClasses/ErrorCorrection/RealignSector.cs
@@ -105,7 +105,7 @@ namespace FloppyControlApp.MyClasses.ErrorCorrection
 
 			
 
-			int markeroffset = sectordata2[indexS1].MarkerPositions;
+			long markeroffset = sectordata2[indexS1].MarkerPositions;
 			byte[] mfmsdest = mfms[sectordata2[indexS1].threadid];
 			var AlignedResult = processing.RealignMFMData4E(ecSettings);
 			//Copy the bitshift correct mfm data back to the large mfms array
